Bug and corrections
First, there was a bug when inserting empty lines in a an empty documents...
The problem was that code did not check the indexes.

Second, we have corrected a bug that prevented Python to return a modified argument.

Third, we have improved the install.py file, which now use a include and a lib path variable toc ompile python.

Furtheremore, we have modified the way the python library is compiled on mac os.