Correction on Unicode
We can now access an Unicode emoji directly through []. In particular, we have added a check_large method, which check through intrinisics instructions if  large Unicode characters are present in a string, in order to trigger a check on emojis.

We have also moved the cleanlibs to the Makefile.in file, in order to have a cleaning that depends on your library selections.