Some corrections and improvments
First, we now use the "allfiles" variables to access the windows through their name instead of looping with instruction sthat should only run in the Main Thread.

Second, we correct a silly bug in ivector, where the "insert" was associated to the wrong method.

Third, we correct a bug in the "ring" container when a value is copied from one ring onto another. The Clear() instruction was missing.

We also do some corrections in the atomic_ring::insert method, which did not take into account the possibility to shift element to the left...

We also correct a bug, in copy. There was a test on the head value that was missing. When this value was NULL, it could crash.

Finally, we have corrected a bug in the way "through" variables are indexed. They were indexed on their numerical name, which could change if a variable was inserted before. We now use the actual name as a string. We also corrected another problem, which was the fact that some "through" types could not be created.

Finally, we have integrated these through variables into the way pre-declared variables are created in the text gui, whcih is part of tamgu.