Harmonization and bug fix
First, we split the an_type set in two different sets: an_type (which now contains only types) and an_flag, which is used for status values of automaton states.
We also add two new methods instead of the ad hoc values used up to now: for an_mandatory (which is now an official value in the an_flag set), we have now ismandatory and removemandatory on the isend and removeend model.

Finally, we fix a bug in TamguFst::process, for endToken, the default value returns when not on a fst end is "found" and not "false" as it was written in the code. This error led to discerpencies between long strings (whith spaces) found in the text and their size.