Updates and corrections
First, the method _bit_scan_forward has a different name on Windows and on Unix platforms. We replace it with a define: bitscanforward (see tamgutypes.h)
Second, there was a bug in the way wstring were tokenized with x_wreading. In some cases, a character too far was read instead of the current one.
Third, the way discrete_distribution is implemented on Windows prevents the use of a vector as input. We needed a more close to metal implementation.
Finally, we add AVXSUPPORT;INTELINTRINSICS to all Windows makefiles.