Improvement and enrichment
First, we now add a read method to map, treemap and we modify the version of vector, which now can takes a complex structure as input.
Second, we improve the way all these strtcures are computed by not using the Push method , which in a thread would create a lock for nothing.
We also correct a minor bug in the parser for "" strings.
We finally correct a bug also in the replace method.