Stable version 1.2020.01.20
First, we switch to a new kind of versionning: V.YYYY.MM.DD, V here stands for version. This is the first Version 1 of Tamgu. This new versionning allows for a simpler way of tracking tamgu different version.

Second, we correct some inconsistencies in jag. When entering a new line, the cursor was too close to the prompt.

Third we correct an issue with find_intel_all and replace_intel. In some cases, we need to realign our string to the previous character, when this previous character matches the first character of the string to be found.