Lisp and other constraints
First, we have improved the tamgu inner editor, which now keeps track of the line which you were editing before running your program.
Second, we have introduced some new operators to the Lisp: in, load, eval etc. We also allows now for regular expressions into lisp expressions.

Third, we now enforce a strict policy for comparisons between different object types. We no longer compare strings and integers as if it was the same object.

We have also corrected some issues with the new indent engine, which is much leaner and faster.

Finally, we have also improved the search of elements by extanlizing from the loop any "if", to reduce branching inefficiency.