Improvement on the Lisp
a) In tamgu shell, we now use a local variable to handle the lisp mode active. This lisp mode allows the user to type in pure lisp command, without the \ and the double quoting in front of atoms.

b) We also add a lisp mode within the interpreter, which can be enables with _lispmode(true);

c) We now manage lisp objects in a pool as strings and vectors.

d) We return an error message when trying to apply computing and comparisons on atoms, except for the equal operator