Docs and Lisp
First, the size method was not declared in the documentation for vectors and maps.
Second, the cond method for Lisp takes sublists, where the first element is a test and the rest the code to execute. The previous implementation reduced the number of elements to only two: one test and one execution?