Modifications
First, we replace a_ppend in Lisp with a_merge, which already existed with the same purpose.

Second, we get rid of Tamgulislong, which is based on list<long> but with terrible performance. We improved function calls by 30%.
We also add Atomref, which generates a new value and increment the reference.
We simplify Local Declaration envirnment by removing the declarations vector, which was redundant and useless.