Large update for speed and bug corrections
First, we have managed to improve significantly the speed of Tamgu by deriving some missing function such Getfloat for some classes. The root version involves the creation and deletion of a temporary object, which consumes useless time. By deriving some Getfloat, Getinteger etc methods, we can directly access the values without any intermediate objects.

Second, there was a specific problem with loading libraries that created objects that were part of the main library, especially the static "methods" container, which stores the matching C++ class methods for tamgu object names. This container was empty for lirbary objects. We now test to see if these structures are empty and use Setidtype to fill them in.

We have replaced the complicated (i-(i << r)) by a "r & 0x3F", which is equivalent in bnimap.h. It is much simpler and much faster.