Correcting a bug in Taskell frame derivation
The issue here was that the frame cretation from which the Taskell data object derived was not fully formed yet.
We postpone now its enrichment to the last stages of construction. As a remindier, Tamgu makes a first pass to create the reference to the different frames and functions then implement the whole stuff in a second pass. Hence, it is possible to call for a function that has been declared after its call.