Updating to Windows
We have enriched tamgusys with mouse control in a Windows Console, which is now on par with Unix systems. We catch mouse events in Window, but generate Unix like strings to keep code as consistant as possible across all platforms. The spreadsheet for instance works fine on Windows, after minor modifications. Most modifications in the spreadsheet were linked to display updates.