BUg and two minor improvements
First, when launching the debugger from the internal editor in tamgu, we need to update idthread:0 (the main  thread) with the thread in which the execution is taking place.
Second, Compilefunction takes also the idthread as argument, to avoid calling GetThreadId for nothing.
Third, we have now an _internals vector, which stores internal information about parsing and compiling speed.