Automatic Resizing
First, we enrich Windows Tamgu internal editor with an automatic resizing mode.
Now, in Windows, it becomes possible to automatically resize the Windows when it is modified.
Second, we add a cllback mechanism to handle resizing through the Tamgusys library. Tamgu exposes now a resizcallback method to add a cllback mechanism to handle resizing. This callback is a two parameters function, which is called when resizing occur. The parameters are the new row size and the column size.