Improvement on trackers in threads
First, we transform trackerslots into an atomic vector, whcih we enrich with a missing backpop method. This method returns the last element of an atomic vector and removes it.
Second, this atomic vector receives now -1 as a default value (zero field in atomic_vector). This enables the removal of trackerslotfilled and simplifies the code as it also enables the remvoval of a specific lock in threadMODE. Since trackerslots is now initialized with -1, a simple test on the output of backpop is enough to detect if a value is available.