Bug with external tokenize rules
The code was not in sync with the more recent development with these rules. parseextern only checks that the rules are ok and then passes the bucket to the regular parserule, which helps keeps all these compiling on the same page.
There was also a problem with checking the arity of a method. We now check the arity with the more specific checkparity that takes the type as an argument.