scanr,foldr
We correct both the Lisp and the taskell version, which did not take into account the way the computing should be donne.
In the case of Taskell, we have now a generic Filterreverse method to apply this operator.
We have also quoted the expressions send to a _map or a _filter in Lisp