A bug correction and a speed improvement
First, there was a bug with the initialisation of a let object in MAIN, when the value was a calculus: let a = 17/3;

Second, we now have a Providelong method implemented. This method helps manages tamgulong objects in a pool.
When a calculus is done in a function call, it is handled in a long object, which was created each time. With a pool of long, the necessary time to handle these calculus has been significantly improved.