Shape
We add a new way to handle matrices with the notion of shape, which has been defined for all vectors.
A shape is a virtual definition of a tensor projected onto a one-dimension vector of values.
We introduce a new way to access elements when this shape is defined: [*i1,i2...] (see Tensor in the updated documentation)

For instance:
fvector fv = [1..12];
fv.shape(3,4);
fv[*1,2] is 7
fv[*-1,1] is [3,7,11]