Bug corrections
First, the choice of "short" for loops within instruction vectors was a poor choice as a large program can take more than 32765 lines.
It is now replaced with a size_t.

Second, there was an issue when launching tamgu -e with a new file. The culprit was the fact that setcode was called too early.