Trimming of code
We remove infomethods and exported from every single instance of class type in Tamgu.

The information are now stored in TamguGlobal in infomethods.
We only keep methods, which keeps track of all methods exported by each type.
We also re-design Setidtype, which is used to initialize an object with its methods. We now test on methods.isEmpty(), which checks if this structure was ever modified.