Corrections and Improvements
We can use _1, _2 etc. as predicate name variables in a rule.

a) We have modified the way the declarations were scanned for variable instances. We now have a specific method: scoot in binmap that replaces the different calls.

b) We add a specific case now in PredicateUnification. If we have not found something to keep but e is a predicate name variable, then we use it as the headpredicate. The reason is that we need to check whether its variable is compatible with the declarations so far.

c) We have a specific name to record these variables in their domain. This name is based on the depth so far, in order to check if two variables belong to the same rule.

d) We have also modified the way depth is incremented. It is only when a new rule is activated that depth is now incremented.

e) We add a new indexation on the second argument of a predicate in the knowledge base.

f) We replace every test: x == "" with x.empty().

g) We rename the different class TamguPredicateKnowledgeBaseFunction into TamguPredicateAction and TamguDependencyAction.

h) We add a checkTypePredicate, which returns for constant predicates their label. It returns a_predicatemethod for TamguPredicateFunction derivations (see TamguPredicateBetween or TamguPredicateSucc for an example).
For actions, such as asserta, assertz we return their label: a_asserta, a_assertz.
For TamguDependencyAction, these labels are modified into a_dependency_asserta etc...
In this way, it is much easier to detect eaxctly which action to apply, releasing many useless tests.

i) We have now a switch case on checkTypePredicate both in PredicateUnification and PredicateEvalue. Again, it avoids many tests on values as before.

j) Stringpredicatekey and Stringpredicatekeysecond now returns a Boolean when a none empty value is stored in the argument.

k) We correct a bug in MethodAddChild. A check on whether the current node was the same as the parent node is now implemented.

l) We introduce a new: Getavariable, which is used in debug methods, to check beforehand if a variable actually exists with this name.

m) There was a bug in TamguJsonCompiler to transform a number into a key. The beginning of the string is at: i - 1.