Minor edit, Alias and Predicate Engine
a) we replace the AFFECTATION keyword (in French) with ASSIGNMENT.

b) We introduce alias, which is a way to write some very simple functions, which can return either a simple value or act as an actual function.

alias test(i) = toto.titi[i];
alias mytest(i,j) = {
   i+=j;
  return i;
}

c) We add the notion of tail recursion to the Predicate Engine. This tail recursion is defined by the user by adding a "#" to the predicate name. This predicate call should be the last in the declaration.

d) We introduce a simpler way to instantiate predicate variable in predicate rules. We now use a TamguPredicateVariableASSIGNEMENT to his end.

e) We get rid of idtype in classes where a constant id already exists

f) The GUI in Windows now only display 50,000 characters at a time.