Minor and major modification
a) First, we rename lockkif into locktamgu

b) We rename Tamgu* call; with Tamgu* function; for consistency.

c) We add a lock to Sqlite calls, to make it possible to use Sqlite queries in threads.

d) In Tamguxml, the call to isConst() was deprecated and led to errors, since a string could be const.

e) We only one single knowledge base, which is now protected for use in threads. You can now execute Prolog requests in threads.

f) We add some new naming convention in Prolog. For instance, _d (where d is a digit) can be used to replace any fact name in the knowledge base.  name#(...), at the end of a rule, can induce a tail recursion.