Some improvements
First, we modify some map instantiation where we used a try/catch. We replace it with a single variable instantiation and a test to see if the variable is NULL.

Second, loadfacts can now loadfacts where one of the arguments is a string without quotes. In this case, it is automatically translated into a string. This feature only works with loadfacts.

Third, the Unique method for vectors is now based on std::set no hmap<type, bool>.

Finally, we add a pool to handle constant strings.