Bug on auto variables
There was a bug on auto variables, when value (see TamguLet or TamguSelf) already contained a const value.
We now replace the test on aNOELEMENT in Put, Putvalue and Setvalue, with a more general test on isConst(), in order to avoid pushing a value into a const, which of course does not do anything.