Mouse on Linux
It is now possible to activate the mouse with VT100 controls on Linux.

We provide -vt100 and -m to activate the mouse in VT100 mode.
On Linux, the default is to have the mouse off in the editor. However, you can either use "-m" on the command line: "tamgu -m", or CTRL-x-m in the editor.
If you want to activate the VT100 mode, you simply need to launch tamgu with -vt100 on the command line.