Code trimming
We got rid of most global variables such as aUNIVERSAL, aTRUE etc. We only keep aNULL, as it is used quite a lot in many part of the code.
This also simplified the exchange of data with an external library.

We also correct a bug, which prevented function variables with the same name as  a Frame variable to be correctly assessed in an _initial function.