Many modifications
The main modification is the use of a much more sophisticated tokenizer, which is now based on a true automaton implementation. We also introduce some mechanism such as tail recursion, which can be inferred or forced in some cases.

This tokenizer works with most encodings: string, u32string, u16string etc.
However, it is implemented on the top of u32string.