Improvement and new features
First, we correct the way the tokenizer is implemented by providing a method to traverse an automaton in a non recursive way. This is specially useful for very large strings, where the recursive version would sometimes create a stack overflow.

Second, we had format strings as in Python: string g = "I={i}";