A correction, a bug and an new feature
First, the way Python handles Unicode string has been modified. PyUnicode_AsUnicode has been deprecated for recent version of the Python API. We use the new method, which is handled through the PY_VERSION_MINOR flag.

Second, in tokenizer_result<STR>::getnext(), the detection was of the CR was done too late, which created some inconsistencies. For instance, the last character before the CR was associated with the next line index.

Third, we add a TamguCheckCompile, which detects all errors in a code. We have also modified the Java API accordingly.