Massive modifications
a) We now replace the errors with an id in a list of error messages
b) returnindextypes now returns the type of values stored in containers.
c) We corrected a bug that prevented fractions to add with themselves.
d) We correction a bug on the negation scope.
e) We now have containers that can contain a specific type of frames