Modifications
First, we normalize many variable names and functions that were too cryptic. For instance: GlobalTamgu is now getGlobalTamgu. The same with the variable "id", which sometimes refers to an idthread, which is now written as idt.

Second, every time we have access to a local globalTamgu object, as in TamguCode, we use it instead on relying on the global variable: globalTamgu.
Third, in the case of the JNI implementation, we now execute the Tamgu code in a thread, which allows to multiple globalTamgu execution. The reason we use threads to execute Tamgu code is that the globalTamgu variable is declared as a thread_local, which means, that even though globalTamgu is a global variable, it can have its own value in a given thread. For instance, in the file codecompile.cxx, we have removed almost every instance of globalTamgu.

Third, We add the method: startwith to detect if a string starts with a specific substring.

Fourth, there was a bug in the way less and lessequal were executed with value NULL values. We now test if these values are NULL beforehand.

Fifth, Constant values such as aNULL, aFALSE are no longer part of globalTamgu but are stored in their own structure, which is never deleted.

Sixth: initialize_local_garbage and clean_from_garbage_position are now part of TamguGlobal et no longer local global variables. It helps running different globalTamgu in parallel. These variables are used to clean objects when using the eval function.

Finally, many methods in globaltamgu.h have now an idx or idglobal