Bug correction
a) We did an important correction to the python library. We now can catch an error when executing a python code with run.

b) We did a lot of corrections to the predicate engine. There was a few bugs, for instance the "cut" operator was not always correctly handled. We also add a new type of function, where a predicate variable can be unified.

c) We add new type: "?_" for functions that can unify predicate variable.

d) We add a specific Execute method in TamguPredicateVariableInstance, which only return a pointer to the variable.

e) We also corrected a bug, when a predicate instance didn't exist, we can now create it on the fly.

f) We improve the error messages when a mismatch occurs in a function call because of incompatible arguments.