New features and bug corrections
a) First, we can now call thread functions from predicates. The only constraint is that after a list of thread calls, you need to add waitonjoined in your predicate rule. We have implemented a new class: TamguCallThreadFromPredicate, which can handle instance variables of type "?_". A thread can unify a Prolog variable.

b) We now protect the access of Setidtype with a lock. Each type class has now a specific classlock variable, which is used to protect Setidtype. The reason is basebin_hash is not thread safe, and the variable "methods" could be altered in different threads.