Feature and bug
First, it is now possible to use: "?_" instead of "predicatevar".
Second, we corrected a few bugs.
One that prevented: loop#(?X,?Y), the tail operator in the predicate engine.
Second, there was an error with the way instances were treated in comparisons. They need to be handled as "auto" variables.