A few corrections
First, when a predicatevar receives a value that is not a predicate, we don't return an error message anymore. We simple push "fail" as a value.
Second, it is now possible to check if a value was initialized in predicatevar with a simple Boolean test such as: if (predvar)...
Third, we now introduce: "?:-" as a synonym to declare a predicatevar.
Fourth, when a regular expression is declared as an argument in a predicate rule, we do not consider this rule as a database entry.