Modification to integrate LispE
LispE is an external library that can be used as a replacement of the current Lisp implementation in Tamgu. This local implementation is both very limited and quite slow while LispE is a true high end Lisp implementation that covers many more field in terms of implementation.
Furthermore, we have implemented a complete equivalence between LispE internal structures and Tamgu's internal structures.