Better stack management
We also correct some potential bugs with bvalues, a static vector that was not always cleaned in function calls.
We also reduce the stack size by getting rid of Tamgudebug elements, which are handled in ThreadStruct.
We reduce the stack whenever we could in PredicateEval for instance, where we got rid of redundant variables and move some heavy structure to the predict class.