A few modifications
First, we move the stack overflow detection to ThreadStruct, in order to keep a much more balanced execution both in the main thread and in other threads.

Second, we correct a bug in ExtractPredicateVariable for maps and vectors. When the value is not a container it must be a predicate variable. Otherwise, we could unify null with []...