Correction in Getvalues
Getvalues when used in initializing a vector in Copyfrom should not fail if a variable in a container is not initialized.
We replace Getvalues duplicate Boolean with a char to be able to differentiate with these 3 potential cases:
a) false: No duplicate
b) true: Duplicate and fail on free variables
c) 2 : Duplicate but no fail on free variables in containers.

We also correct a bug when some elements with free variables were considered facts. We now check if there are free variables in internal representations such as term or containers.