Some corrections
For lispe, we now have a better use of @@ at the limits. 0 is now the last element of a list.

We add a header list management for the specific case of CURLOPT_HTTPHEADER in libcurl. We need objects with a specific type to handle values for this method.

We also correct a bug when a global variable is still not available when launching a thread.